Implements CSFVectorFunction::enumerate() member functions
The commit brings two versions of enumerate() for the CFSVectorFunction
class. The first one performs a global enumeration which implies MPI
communication. The other version just performs local operations and
receives an offset for starting the indexing.

We extend VectorFunctionBasicTest to test enumerating.

In order to get enumerate working for P1VectorFunction<int>, the commit
makes enumerate() with offset public in VertexDoFFunction.