Extending capabilities of VTKOutput
This commit changes VTKOutput so that it export also functions of the
following kind:
- P1Function< int32_t >
- P1Function< int64_t >
and not only P1Function< double >.

For this we replace the std::vector< P1Function< real_t > > in the class
by an object of type FunctionMultiStore. This new class allows to store
multiple functions from the same family, which can but need not differ in
their value types.