Changes to typedefs/using statements in some function classes
The FunctionWrapper class no longer typedefs FunctionType to its
func_t argument. IMHO that was pretty useless. Instead we get a
```
template< typename VType >
using WrappedFuncKind =
typename WrappedFuncType::template FunctionType< VType >;
```
For this to work functions the get wrapped need to provide a templated
FunctionType. So this is added to
- DGFunction
- and all already exisiting children of the BlockFunction class.