Simplifies implementation of BlockFunction::getNumberOfLocalDoFs()
The previous implementation relied on directly using the templated
free function numberOfLocalDoFs(). In order to call the correct one
a switch statement was used based on the FunctionKind of the
corresponding sub-function.

Consequently we would need to update this code piece for every new
type of sub-function that we want to use.

The new implementation is directly using
GenericFunction::getNumberOfLocalDoFs().