Implements PETSc vector <-> function conversion for GenericFunction
The commit introduces implementations of
- createVectorFromFunction()
- createFunctionFromVector()
for functions of type GenericFunction< PetscReal >.

The two free functions reside in the new file GenericFunctionPetsc.hpp.
Because of the peculiarities of GenericFunctions these free functions
are, however, only wrappers that call the corresponding one of the two
new member functions
- GenericFunction::toVector()
- GenericFunction::fromVector()
As usual these are pure virtual and the actual work happens inside
the member functions of the FunctionWrapper class.

We extend the FunctionWrapperTest to check compilation and execution
for various kinds of wrapped functions.

Note that due to instantiation requirements the commit also implements
- createVectorFromFunction()
- createFunctionFromVector()
for functions of type DGFunction. However, these implementations are
pseudo only and will abort.