Reworks numberOfLocal/GlobalDoFs for Block/GenericFunction
We now treat the special case of having a BlockFunction or GenericFunction
in the templated function numberOfLocal/GlobalDoFs directly using a
constexpr-if. Makes it much easier for the compiler, which otherwise
sometimes had problems finding the correct variant.

For this to work BlockFunction::ValueType was renamed to
BlockFunction::valueType like in the other functions and a similar
typedef was added to GenericFunction.