Adds two "full stokes" forms for P2 elements
Commit adds
* p2_full_stokescc_affine_q3.[ch]pp
* p2_full_stokesvar_blending_q3.[ch]pp

While the first one ist more for internal comparison purposes (for constant
coefficients the epsilon part of the full stokes would reduce to the vector
laplacian) the latter one is the one we need for the anelastic mantle
convection models.

Commit extends the HytegVsFenicsFormTest for the two new forms.