Fixes bug in BlockOperatorBasicTest
runTest for kind=1 was erroneously using P2ConstantLaplaceOperator instead
of P2ConstantVectorLaplaceOperator for the (0,0) block. This led to a memory
problem when calling P2Function::isDummy(), and, unfortunately, did not
necessarily crash.

Additionally commit adds a reminder that the second invocation of runTest
is actually unclean, since we work with a P2P1TaylorHoodBlockFunction,
but should use a P1P1StokesBlockFunction.