Began implementation of VolumeDoFFunction.
The VolumeDoFFunction is supposed to allocate DoFs in each volume (faces in 2D and cells in 3D)
and allow for a variable number of DoFs per macro-primitive. Also it supports SoA and AoS memory
layouts of the DoFs. Communication shall only be performed from volume-primitive to volume-primitive.
Interfaces are not used at all. Ghost-layers are added directly to the volume-primitives.

This is meant to be the basic data structure for the DG implementation.