Started implementation of the new DGFunction.
Once finished this shall deprecate the previous implementation.

It comes with a DGBasisInfo interface that is going to have generated implementations
to assign basis functions to the DoFs. DGBasisInfos must implement an evaluate method
and evaluation of the linear form \int_T f * phi_i dx to assemble right-hand sides
and to project scalar functions into the space.