Commit formally makes VectorLaplaceOperator smoothable
The VectorLaplaceOperator class now publically inherits from
- GSSmoothable
- GSBackwardsSmoothable
- SORSmoothable
- SORBackwardsSmoothable
and implements the associated pure virtual methods. The
implementation in turn checks that the diagonal sub-operators
have the corresponding interface and delegates work to them.

Note: Of course a P2ElementwiseVectorLaplaceOperator is still
not truely SORSmoothable. However, that will only be recognised
on the sub-operator level.