Makes ChebyshevSmoother instantiable for "VectorOperators"
Using constexpr if, we can make the ChebyshevSmoother formally
be instantiable for operators whose SourceFunction is a child
of CSFVectorFunction.

The problem lies with the assertion calling getMaxMagnitude,
which is currently not provided by CSFVectorFunction.

Note that this is only a formal change. Since we have no
vector operators that return their inverse "block" diagonal, and,
thus, none that inherit from OperatorWithInverseDiagonal,
the ChebyshevSmoother is still not usable with these kind of
operators.