Step towards disentangling StokesOperators and Solvers/Preconditioners
Most (I probably overlooked at least one) composite Stokes operators now
provide a getA() member function to access to velocity block. Commit
starts replacing explicit accesses to A by getA() in solvers and
preconditioners. By far not complete, yet.

Actually the A block gets removed by the commit as the composite Stokes
operators now make use of VectorLaplaceOperators instead.

We also add to many a div and divT operator that is a vector-to-scalar
resp. scalar-to-vector operator. The old scalar sub-operators are kept
for the moment, because removing them implies refactoring the whole
Uzawa smoother.

Additionally the elementwise Stokes operators cause a problem here due to
the fact that we cannot call computeAndStoreLocalElementMatrices on
div or divT and it is not straightforward to hand sub-operators back from
these template classes.

Implements toMatrix() method missing in some operators:
- P1ScalarToP1VectorOperator
- P1VectorToP1ScalarOperator
- P2ScalarToP2VectorOperator
- P2VectorToP2ScalarOperator