Resolves issue #176
Part 1:
The commit introduces more generic versions of scalar-to-vector and
vector-to-scalar operators. This allows to remove the individual
implementation for different function space combinations.

PArt2:
The two classes ScalarToVectorOperator and VectorToScalarOperator now
provide templated member functions getSubOperator. The template parameter
determines the index {0,1,2} of the desired sub-operator.

The new accessors allow to continue refactoring the elementwise Stokes
operators.