Starts a new terraneo module
As discussed in the last developer meeting it is time, in light of soon to
come extensions, to start with a separate terraneo module. This is intended
to house code specifically addressing the needs of running mantle convection
models with HyTeG.

This commit starts with the new module and relocates the
SphericalHarmonicsTool into it.