Updates MeshGenSphericalShell
The mesh generator for thick spherical shells now also flags primitives
using the MeshInfo::hollowFlag enum, like that for the full annulus does
already. This means that vertices, edges and faces on the outer surface
are marked as flagOuterBoundary, those on the inner as flagInnerBoundary
and all others and of course cells as flagInterior.

The new HollowBodyUIDTest checks that this works correctly for both types
of hollow bodies.