Further work on plate boundary conditions
Commit reduces verbosity of plate information handling code and starts
an attempt to make handling of cases when we cannot determine a plate
associated with a point more transparent.

The file PlateNotFoundHandlers.hpp contains two classes
- DefaultPlateNotFoundHandler
- StatisticsPlateNotFoundHandler
The first one just keeps doing, what we did so far, which is set the
point velocity to zero, if we have no plate information. The latter
does the same, but remembers the point's coordinates and the age stage
to allow generation of a report in the end.