Adds (de)serialisation for dense Eigen::Matrix objects
The idea is to be able to MPI-communicate dense Eigen::Matrix objects
via waLBerla functions. For this they should be serialisable to a buffer
and deserialisable from a buffer. Not too difficult actually, so the
commit implements this for a general dense Eigen::Matrix of arbitrary
dimensions. Commit also adds a test to check this and some other basic
Eigen functionality. So we know interfacing works.