Fixes issue #183
This patch should fix issue #183 by making the setup computations for the
blending in the IcosahedralShellMap more robust for large radii.

For finding the two vertices an a ray from the origin we no longer only
consider the norm of the cross-product of the position vectors, but scale
this be their lengths.

Additionally classifyTets() no longer uses a plain tolerance when testing
for equality of a point radius with the radius of the outer shell, but
checks that the difference is smaller than one permille of the layer
thickness.

Additionally the commit fixes some inline comments and debug output.