Reduce code-duplication in evaluate() methods
The commit adds two new free-functions
- findFaceIDForPointIn2D()
- findCellIDForPointIn3D()
These do precisely what previously was implemented in the evaluate()
methods of the P[12]Function and DGFunction classes, i.e. trying to
determine to which macro-face/cell a point in 2D/3D belongs or
alternatively extrapolating inside a circle/ball.

In the spirit of C++ Core Guidelines F.20/F.21 the functions return a
tuple of bool and PrimitiveID.