Extend GeometryMap concept
The commit adds a virtual method GeometryMap::verifyPointPairing().
This will always return true. In the AnnulusMap and IcosahedralShellMap
classes we override it to check that the corresponding (local) map
maps the given coordinates on the computational domain to those on the
physical domain. For this we can use a tight tolerance.

The new method is used by the free-function (in BlendingHelpers.hpp)
mapFromPhysicalToComputationalDomain2D() to check that the face we
identify to belong to the point is the correct one.