Modify GeometryMap::verifyPointPairing()
We make GeometryMap::verifyPointPairing() a pure virtual method and
implement it in all maps individually. Those that are defined globally

- IdentityMap.hpp
- AffineMap2D.hpp
- AffineMap3D.hpp

simply return true. The others will check whether the distance between
the point in the physical domain and the point in the computational
domain when the latter is mapped into the physical domain is shorter
than the value of GeometryMap::defaultThresholdForPointComparison.