Fix bug in evaluateGradient()
The implementation of P[12]Function::evaluateGradient() by accident
created a copy of the macro face on which to evaluate the gradient.
While primitives can be copy constructed, the attached data, or more
precisely the map with the pointers to the data, are not copied.
So that could not work. Commit changes that to creating a reference
instead, as intended.