Fix P2EvaluateTest
We cannot expect to get accurate results when we use the annulus map
and use an arbitrary 2nd order polynomial on the physical domain, as
it will not map to a 2nd order polynomial on the computational domain.
We use a special function now that results in something that can be
represented exactly on the computational domain.