New shot at deallocation. For now, functions are still handles. This means that copying a function does not produce a deep copy of the allocated memory but yields another object handle pointing to the same memory. We might change that in the future.
This commit tackles the problem of memory deallocation when functions go out of scope. Somehow we must know whether the allocated memory is still referenced via PrimitiveDataIDs. Adding reference counters to the IDs themselves would be elegant but this does not work since copies are stored in various places and not only in the functions themselves. While their reference count might not be zero yet, deletion might still be safe (they are for instance captured by lambdas in the PrimitiveStorage).

So I added reference counting to the functions (here VertexDoFFunction). Copies are still shallow, but once all VertexDoFFunctions with the same reference counter are deleted, the memory is freed. Let's see if that is a step into the right direction...