Refactor GeometryMap::verifyPointPairing()
The previous approach was too closely tied to the use of that method
inside functions mapFromPhysicalToComputationalDomain[23]D. There it
is clear that for a global homeomorphism we have no problem and e.g.
the IdentityMap can always return true.

However, if called with an arbitrary pair of coordinates from the
computational and physical domain there is no guarantue that those
are mapped onto each other even for the IdentityMap.

We implement the method now in GeometryMap itself. Children could
still re-implement special version, if necessary.