Fixes problem associated with issue #187
Apparently in the plate data polygons can have "double point", i.e. in
the list of points two consecutive entries may be identical. This caused
a division-by-zero error in intersectPointWithLine() as the length of
that polygon segment is then zero.

Commit adds a WALBERLA_ASSERT to intersectPointWithLine() to catch this
problem and avoids it in findPlateAndDistance(), by changing the logic
to first check whether lineStart and lineEnd are identical and not after
the computation.