Fix issue with Matrix class
Eigen does not support creating Eigen::Matrix< M, N, RowMajor > with
number of columns being N = 1. However, the new DG forms work with
column-vectors of the form Matrix< real_t, M, 1 >.

We use a private static constexpr function now to make sure that
for N = 1 we use Eigen::ColMajor as storage type. Note that this
does not create an issue with Matrix::data() as we basically bave
a vector then.