Implement MeshInfo::meshThinSphericalShell()
The commit adds a new function MeshInfo::meshThinSphericalShell()
which generates a MeshInfo object representing an icosahedral mesh
for a thin spherical shell of given radius.

We extend the MeshGenTest test and the show_mesh app to treat also
this feature. Additionally we add a ThinShellMap class to perform
the associated blending operation.