Add a SphericalElementFormMass
This form computes the element mass matrix for the elements derived
by Baumgardner and Frederickson for spherical triangles.

The elements have three DoFs per element associated with the vertices.
We can, thus, use a VertexDoFFunction for them, which we alias for this
purpose to SphericalElementFunction.

Commit extends MassOperatorTest to check that we can correctly compute
the surface area of a thin spherical shell with given radius using a
P1ElementOperator< SphericalElementFormMass >.