Fixes the high velocities due to wrong rec tree
This change avoids having large velocities due to the creation of a
wrong reconstruction tree. Some plates have two reconstruction poles
at the same time, the plate reconstruction tree was set to get the
older plate ID instead of the current one.
Now the code is set to create the reconstruction tree with the age
provided by the user.