Disentangle VectorFunction dimension from hasGlobalCells()
The commit allows to pass an argument vectorDim to the constructors
of the P1[12]VectorFunction classes. This must have a value from
{0,2,3}.

In the case vectorDim == 0 the decision on the dimension of the vector
field is based on whether the PrimitiveStorage contains cells or not.
Zero is the default argument value to preserve previous behaviour.

In case of a value of 2 or 3 this will become the dimension of the
vector field. This allows us to have 3D vector fields also on 2D
surfaces.