Remove assembly query methods from non-generated forms
This commit is step 1 in resolving issue #192. We remove the following
methods
- assemble2D()
- assemble3D()
- assembly2DDefined()
- assembly3DDefined()
from all form classes besides
- the ones generated with HFG
- the base class Form itself
In the latter we implement them as virtual functions instead of being
pure virtual as before.

Once we have regenerated all HFG forms we can also remove the methods
from the Form class itself.

With this commit we also remove the associated checks from the operators
- EdgeDoFOperator
- EdgeDoFToVertexDoFOperator
- VertexToEdgeDoFOperator