Point in polygon in 3D using the library boost
It implements the point in polygon in the surface of the sphere
using the libary boost. To do so, it uses the coordinates in
"spherical-equatorial", lon lat as used in geophysics (see boost page).
Thus the plates boundaries are now stored directly with lon lat,
they are not converted to the unit sphere.

It also calculates the distance from the point to the polygon using
the boost library.

This implementation reduces the amount of operations to locate the point
into a plate and removes the dependency of the module plates
to the library CGAL.
It also reduces the amount of points without a plate, almost to none.
The only times we have a point without a plate, is due to the gaps that
the input file provides.

(cherry picked from commit 163da0bc215568ccfd4af74fc7cd2d100474187e)