Cleanup and change of internal coordinate system
General cleanup of functions that were not used anymore due to
the update to point-in-polygon in 3D using the boost library.

Change also to the interal coordinate system to the usage of the
coordinates lon lat. This change is due to the boost library.
It preforms the operations in the surface of a sphere in such
coordinate system.
Also also to improve usability of the code.
The user doesn't have to change the input coordinates to the unit sphere
anymore. Now, palates can be given any xyz, independant of the radius.

The output is still the velocity of the point in xyz in the sphere of
EarthRadius in m/s.

(cherry picked from commit 9bce1d330c543bd2c3bc5df82c7b45968f0d8710)