Re-add functionality to export DGFunction_old
Running apps/dg0_transport I notices that while one could still add a
DGDunction_old to a VTKOutput object, the added function was no longer
exported.

I know that we plan to get rid of the old implementation stuff, but for
this I would like to be able to compare old vs. new. Also the behaviour
was a little disconcerting.

Thus, I re-added invocation of VTKFaceDoFWriter::write(), if functions
of type FaceDoFFunction_old are present.