Convert DGUpwindOperator to DG0P1UpwindOperator
While the old DGUpwindOperator was templated with a velocity type it
actually only supported using velocity fields in 2D with P1Function
components.

This commit de-templates and renames the operator. It also changes it
to internally store the velocity field as a P1VectorFunction.

The apps using the operator and the Upwind2DCircularConvectionTest are
modified accordingly.