Replace VertexDoFFunction::integrateDG by free-function.
This commit replaces the member function integrateDG of the VertexDoFFunction
class by the free-function P1IntegrateDG(). On the one hand, this functionality
is too special to be placed in the class and better off in a stand-alone
function. On the other hand this is an intermediate step in removing
DGFunction_old, as it makes porting it to P0Function easier.