Extend BasicBlendingMapTest
The commit adds the remaining currently implemented blending maps to the
tests. The test now also checks that we can serialize and deserialize the
maps and that we at least obtain the correct type of map after deseriali-
zation. Note that TomomakMap and ThinShellMap currently do not support
(de)serialization.