Make HyTeG aware of system's byte order
Commit adds a new macro HYTEG_ARCH_ENDIANESS to HytegDefinitions that
is set by CMake to the value of CMAKE_CXX_BYTE_ORDER. We also add a query
method systemEndianess() to BuildInfo.

We currently use the endianess information to add setting the byte_order
attribute of the VTKFile element for VTU files.