Fixes the zero/nan velocities for old plate recon.
It fixes the zero or nan velocities when the time we ask to calculate
the velocity is an older limit of the age of the  finite rotations
given to plates (included in the rotational file).

As a default, we calculate the euler pole using two finite rotations
at (time, time+1). With this fix, when the time we are want the velocities
is the last listed in the finite rotations in the rotational for a given
plate, we use the approximation of (time-1, time).

This is in line with what gplates offers. And obtains the same results
as calculating the velocities using pygplates for conflicting times and
platesID.