Extend BasicADIOSTest to really do something [skip ci]
The test now sets up a P1Function on a single triangle macro mesh and
exports the complete data held by the macro face via ADIOS2. It then reads
the data in again to setup a cloned function and compares it to the original.
This is done twice once for the BP5 engine (so file IO) and then with the
"inline" engine. The functions's name is attached as an attribute to the
data.