Fix issue #221
This commit fixes issue #221, i.e. PETScLUSolver::solve() will again
copy the boundary condition information from the vector x of unknowns
to the numerator object. This happens before the assembly and factorisation.

Note, however, that we currently only do this, when the numerator object
was created internally. We do not touch it, when it got passed from the
outside.