Remove useVTKQuadratic[Triangle|Tetra]_ from VTKOutput class
The commit removes the two attributes and associated getter/setter methods
from the VTKOutput class. They had been introduced as part of the transition
to write P2Elements as such and not as P1Elements on a refined mesh. With
this we also remove support for the old approach and the associated tests.