Start implementation of an AdiosWriter class and associated test
The commit introduces two new classes
- AdiosWriter
- AdiosWriterForP1
The first one serves as the app interface in the same fashion as VTKOutput
does and is another child of FEFunctionWriter.

The second class handles the details of exporting P1Functions and
P1VectorFunctions and is the one actually interfacing with the ADIOS2
library.

Reuse of functionality from the refactored VTKMeshWriter is accomplished
via the AdiosWriter::StreamAccessBuffer class that acts as a wrapper for
Adios' Variable<T>::Span objects.

The current draft is working, at least for static problems in 2D and 3D.