Refactor VTKMeshWriter class
The methods of the VTKMeshWriter class are now independent of VTKOutput,
i.e. we pass information directly instead of passing a reference to the
complete VTKOutput object.

The two methods
- writePointsForMicroVertices()
- writePointsForMicroEdges()
are now templated. The template argument is the stream to which they write
the node coordinates. The other VTKWriters now pass a VTKStreamWriter instead
of an std::ostream and perform the VTKStreamWriter::toStream() call themselves.

The commit factors out the writing of the element <-> node association
from the methods
- writeCells2D()
- writeCells3D()
- writeConnectivityP2Triangles()
- writeConnectivityP2Tetrahedrons()
and places it into the new templated methods
- writeElementNodeAssociationP1Triangles()
- writeElementNodeAssociationP2Triangles()
- writeElementNodeAssociationP1Tetrahedrons()
- writeElementNodeAssociationP2Tetrahedrons()
As aboved the template argument is the type of stream to which the data
get written.