Add macros for activated features to Doxygen's PREDEFINED list
So far the Doxygen documentation is decoupled from the features we
acitvate or deactivate for a build. So, for example, Doxygen is unaware
whether WALBERLA_BUILD_WITH_MPI is defined or not.

This commit adds macros for activated features to Doxygen's PREDEFINED
list, so that the correct documentation is generated.

Note: I find my solution in CMakeLists.txt akward. However, given my
limited CMake expertise I was unable to come up with something more
elegant. If you do, please change it :)