Change FEFunctionWriter to use CRTP
The previous inheritance concept, with FEFunctionWriter as base class for
the concrete (AdiosWriter and VTKOutput) child classes was faulty. The
problem was that the add() method is templated and templated methods are
currently not allowed to be virtual, too.

So with this commit we change to CRTP. Not sure how much sense this makes,
but it will at least enforce a minimal level of consistency in the methods
offered by the "child" classes.