Add BlockFunction::get[Min|Max]Level()
For this we also add the corresponding methods to the GenericFunction
and FunctionWrapper classes.

Additionally the commit removes the const from the functionName_ member
of BlockFunction to allow generation of the implicit copy assignment
operator. Thus, the corresponding part of the FunctionRuleOfThree test
then also works for a P2P1TaylorHoodFunction.