Re-implement AdiosCheckPointExporter w/o using adios2::Span
The previous implementation made use of adios2::Span for concertedly handling
data of the component functions of a vector function as well as the VertexDoF
and EdgeDoF components of a P2 function makes data import inefficient. This
leads to a significantly smaller number of variables one needs to deal with
in the checkpoint, e.g. one adios2::Variables for a P2VectorFunction in 3D as
opposed to six.

However, different from Put() the Get() in Adios2 does not support spans.
Also, there seems to be no support for reading only parts of a process local
variable in Adios2. Thus, we would need to allocate our own temporary buffer,
which seems not ideal.

The re-implemented version treats all component functions as separate
variables. In the end this even leads to a significant reduction in lines
of code and complexity.