Make names of adios2::IO objects in AdiosWriters more generic
We unfortunately need to create one adios2::IO object for each output
level and for each function family. Previously their names were given
by the full name of the respective output including its path. This
is inconvenient for setting parameters via a config file.

With this commit we change the names of the IO objects to be
- AdiosWriterP1-level<level number>
- AdiosWriterP2-level<level number>