Allow float16 types for the P1 function space.
Template and explicit specialize generated kernels <add, apply, assign, communicate> in the P1 function space (the other kernels are not templated and can therefore not be used for float16).
Change the C++ standard to cxx23 in for all files that use float16, when WALBERLA_BUILD_WITH_HALF_PRECISION_SUPPORT is enabled.
Add tests that float16 works in general and activate them in the CI.
Update walberla to support float16.