Refactored temperature initialization.
Removed the class templates in favor of free functions that construct lambdas that can later be passed to the interpolate methods of HyTeG's FE functions.

This is more flexible and extensible (the functions can be nested - which is already done for two of the implemented cases) and reduces coupling.

Also adapted documentation.