Introducing RadialShellData.
Added class RadialShellData to store DoFs of P1 and P2 functions organized into radial shell point clouds.

Also, some edits to the radial profile tools. Modifies the computation of the number of shells - I think this was not 100% correct before, but I might be one. Needs some more testing - because this also affects the RadialProfile computations.