Update connectivity computations in VTKMeshWriter
The writeElementNodeAssociation* methods no longer use a special integer
type CellType / CellIdx_t etc. when computing the connectivity information.
Instead they just use HyTeG's uint_t.

The conversion of uint_t data to whatever integer datatype is to be used
in the corresponding VTKOutput or AdiosWriter is left to be handled by the
corresponding data output stream. The VTKStreamWriter was prepared to do
this or did this already anyhow. The AdiosWriter::StreamAccessBuffer now
can also do this.

Note: There currently is no checking that data ranges will be sufficiently
large when doing the conversion.