Allow selecting signedness of connectivity info in BP files
The commit adds a compile time option HYTEG_USE_SIGNED_INT_FOR_ADIOS2.
By default this is OFF, but will be forced to ON when compiling with
HYTEG_TERRANEO_MODULE=ON.

Depending on the setting we define the macro ADIOS2_PARAVIEW_INT_TYPE
to be uint64_t or int4_t. The AdiosWriters will use this macro to
select the data type of the connectivity information in the BP file.