Add solverLoop for Uzawa type multigrid solver approach
A solver loop is added to perform multiple V-Cycles using the Uzawa type
multigrid Solver for the Stokes system.
The solver loop will end when either an absolute- or  relative residual threshold is reached or
a maximum reduction rate of 0.5 between the current residual estimation and the previous residual estimation is exceeded.
A user can define the mentiond maximum number of coarse grid iterations, absolute- and relative residual threshold within the
corresponding  parameters file.

Minor cleanup and fix