Add a prependHyTeGMeshDir() function
As discussed internally we would like to make opening "internal" mesh
files in tests and tutorials resilient against re-location of source
code files.

This commit introduces a HYTEG_MESH_DIR macro that stores the absolute
path to "data/mesh" inside the build directory. Additionally it adds
a free function prependHyTeGMeshDir() that can be used to prepend this
path to a filename or relative path.