Extend GmshReaderForMSH41 to set boundaryFlags from physical tags
MeshInfo::fromGmshFile now has an optional flag importPhysicalTags. By
default this is true. If true, then the GmshReader class for the MSH4.1
format will read the physical tags of the entities defined in the meshfile
and use these to set the boundary flags of the vertices, edges, faces and
cells it imports and derives.

For this to work all entities, used in the mesh, must be present in the
"Entities" section of the file and have been associated a physical tag
via the "Physical Curve" etc. commands of Gmsh.

If the flag is false, the "Entities" section need not be present and
all primitives will be flagged as 0.