Add boolean parameter excludeDGTypeFunctions to syncRegisteredFunctions
This mandatory argument allows the caller do decide, wether DG-type
function in the FEFunctionRegistry should be included into the
synchronisation or not (like in the case of VTKOutput and AdiosWriter).

Note that synchronisation for DG-type functions is not implemented,
in this function, yet. Thus, it will abort, if DG-type functions are
present and excludeDGTypeFunctions is set to true.