Update CheckpointRestoreTest(s)
The pipeline runs various versions of the CheckpointRestoreTest. This commit
better separates them by making the test append the number of MPI processes
to the filenames.

It also adds FIXTURES to make CMake aware of the dependencies between the
"normal" jobs and those that only import files, which, of course, need to
have been produced beforehand.

Additionally we now enforce for the import-only variant that the user, either
in the parameter file or as a command-line override sets the name of the
checkpoint file to import data from.

Finally we add different variants of the import-only job.

This should fix issue #280.