Update some tests for the new 3D version 4.1 mesh-files
The two new 3D mesh-files, tet_1el.msh and pyramid_2el.msh generated
with a current Gmsh version and exported in MSH4.1 format use a different
sequence of the node indices when describing the 4-node-tet element(s)
than the previouis MSH2.2 files.

This has consequences on the sequence of primitives in the PrimitiveStorage
and also on the discretisation, due to a different choice of diagonal
when refining. Hence, some tests failed, see e.g. issue #279.

This commit adds variants of the two files, still in MSH4.1 format, but
using the previous sequence of node indices. Some of the failing tests,
especially those that have a built-in dependency on the sequence and
meaning of specific primitives in the PrimitiveStorage, are updated to
use the variant mesh-files. The remaining tests have their thresholds
adapted.