Add a GmshExportTest
The commit adds a test to check that the functionality for exporting
refined meshes compiles and runs through for different 2D and 3D cases.

Optionally the test can verify the generated meshes by parsing them
with Gmsh. To allow this CMake now checks for the presence of a gmsh
executable.

The commit also makes exportRefinedMesh() less verbose by changing messages
from PROGRESS to DETAIL.